refactor: split compiler.ts into distinct owner stages
Split the 1,601-line monolithic compiler.ts into 5 focused stage modules:

- source-parser.ts (SourceParser): markdown text → SourceIR
  (headings, raw metadata, raw route blocks, anchors)
- graph-compiler.ts (GraphCompiler): SourceIR → pattern/route/relation graphs
- index-projector.ts (IndexProjector): SourceIR + graphs →
  lexicon, index map, compiled nodes, indexes
- validation-runner.ts (ValidationRunner): snapshot candidate →
  validation findings
- query-helpers.ts: query scoring/matching functions consumed by
  QueryEngine

compiler.ts remains as the SnapshotAssembler orchestrator (144 lines),
wiring stages together and re-exporting query helpers for backwards
compatibility.

Public compileFpfSource() API is unchanged. All 39 existing tests
pass without modification.

Closes #1
